requirements.txt
gymnasium==0.29.1
stable-baselines3==2.1.0
numpy>=1.24.0
pyyaml>=6.0
requests>=2.28.0
streamlit>=1.28.0
matplotlib>=3.7.0
pandas>=2.0.0
tensorboard>=2.14.0
